Returned canInitiateJetpackRestConnection to the view model
